testing/requirements.txt
langchain-core
langgraph
pydantic
langchain-google-genai
python-dotenv
langchain
langchain-community
pymysql